fourth last commit simple design

+++ SaitZaObuvkiComfort/SaitZaObuvkiComfort/Data/Product.cs
@@ -8,7 +8,7 @@ namespace SaitZaObuvkiComfort.Data
         public string Name { get; set; }
         public int CategoryId { get; set; }
         public Category Categories { get; set; }
-        public decimal Size { get; set; }
+        public int Size { get; set; }
         [Column(TypeName ="decimal(10,2)")]
         public decimal Price { get; set; }
         public string PhotoURL {  get; set; }
